Only consider same version for save transfers
This change will prevent savegame-transfer from transferring to a
trunk version that does not target the same stable version. In other
words, a trunk game on 0.22-a0 will not be transferred to 0.23-a0.
It will stay at the latest available trunk version targeting 0.22.

This is combined with the limitation on savegame compatibility.
Therefore, savegame-transfer will now transfer the selected
character to the latest version that has the same "major version"
(savegame compat) and the same game version (0.23, for example).

WARNING: Make sure that you never delete the latest trunk version
available for any given stable version (0.23). If you follow this
rule, players will always be able to transfer to the best trunk
version available for their stable version, even long after that
stable version is stale.